requirements.txt
Django==4.0.4
djangorestframework==3.13.1
google-api-python-client==2.44.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.1
requests==2.27.1
colorama